+++ src/main/java/io/github/mapepire_ibmi/types/GetTraceDataResult.java
@@ -37,7 +37,7 @@ public class GetTraceDataResult extends ServerResponse {
      * @param traceData       The trace data.
      * @param jtOpenTraceData The JTOpen trace data.
      */
-    public GetTraceDataResult(String id, boolean success, String error, int sqlRc, String sqlState, int executionTime,
+    public GetTraceDataResult(String id, boolean success, String error, int sqlRc, String sqlState, long executionTime,
             String traceData, String jtOpenTraceData) {
         super(id, success, error, sqlRc, sqlState, executionTime);
         this.traceData = traceData;
